+++ c_src/keyio.c
@@ -12,15 +12,17 @@ int ioctl_keyboard(int fd, int grab) {
   return ioctl(fd, EVIOCGRAB, grab);
 }
 
-// Acquire a filedescriptor as a uinput keyboard
+// Acquire a file descriptor as a uinput keyboard
 int acquire_uinput_keysink(int fd, char *name, int vendor, int product, int version) {
 
   // Designate fd as a keyboard of all keys
   ioctl(fd, UI_SET_EVBIT, EV_KEY);
   int i;
-  for (i=0; i < 256; i++) {
+  for (i=0; i < 0x2ff; i++) {
     ioctl(fd, UI_SET_KEYBIT, i);
   }
+  // Enable key repeat for output keyboard
+  ioctl(fd, UI_SET_EVBIT, EV_REP);
 
   // Set the vendor details
   struct uinput_setup usetup;