+++ src/KMonad/Args/Types.hs
@@ -16,13 +16,15 @@ module KMonad.Args.Types
 
     -- * $but
   , DefButton(..)
+  , ImplArnd(..)
 
     -- * $tls
   , DefSetting(..)
   , DefSettings
   , DefAlias
+  , DefLayerSetting(..)
   , DefLayer(..)
-  , DefSrc
+  , DefSrc(..)
   , KExpr(..)
 
     -- * $defio
@@ -32,6 +34,8 @@ module KMonad.Args.Types
     -- * $lenses
   , AsKExpr(..)
   , AsDefSetting(..)
+  , HasDefSrc(..)
+  , AsDefLayerSetting(..)
 ) where
 
 
@@ -71,7 +75,11 @@ data DefButton
   | KAroundNext DefButton                  -- ^ Surround a future button
   | KAroundNextSingle DefButton            -- ^ Surround a future button
   | KMultiTap [(Int, DefButton)] DefButton -- ^ Do things depending on tap-count
+  | KStepped [DefButton]                   -- ^ Do different things, one-by-one
   | KAround DefButton DefButton            -- ^ Wrap 1 button around another
+  | KAroundOnly DefButton DefButton        -- ^ Wrap 1 button only around another
+  | KAroundWhenAlone DefButton DefButton   -- ^ Wrap 1 button around another when it's "alone"
+  | KAroundImplicit DefButton DefButton    -- ^ Wrap 1 button around another
   | KAroundNextTimeout Int DefButton DefButton
   | KTapMacro [DefButton] (Maybe Int)
     -- ^ Sequence of buttons to tap, possible delay between each press
@@ -87,13 +95,22 @@ data DefButton
   | KBeforeAfterNext DefButton DefButton   -- ^ Surround a future button in a before and after tap
   | KTrans                                 -- ^ Transparent button that does nothing
   | KBlock                                 -- ^ Button that catches event
-  deriving Show
+  deriving (Show, Eq, Typeable, Data)
 
+instance Plated DefButton
+
+-- | Possible values for implicit around
+data ImplArnd
+  = IADisabled
+  | IAAround
+  | IAAroundOnly
+  | IAAroundWhenAlone
+  deriving (Show, Eq)
 
 --------------------------------------------------------------------------------
 -- $cfg
 --
--- The Cfg token that can be extracted from a config-text without ever enterring
+-- The Cfg token that can be extracted from a config-text without ever entering
 -- IO. This will then directly be translated to a DaemonCfg
 --
 
@@ -106,6 +123,7 @@ data CfgToken = CfgToken
   , _fstL  :: LayerTag                          -- ^ Name of initial layer
   , _flt   :: Bool                              -- ^ How to deal with unhandled events
   , _allow :: Bool                              -- ^ Whether to allow shell commands
+  , _ksd   :: Maybe Int                         -- ^ Output delay between keys
   }
 makeClassy ''CfgToken
 
@@ -116,18 +134,31 @@ makeClassy ''CfgToken
 -- A collection of all the different top-level statements possible in a config
 -- file.
 
--- | A list of keycodes describing the ordering of all the other layers
-type DefSrc = [Keycode]
+-- | A list of keycodes describing the ordering used by all other layers
+-- | which is associated with a name.
+data DefSrc = DefSrc
+  { _srcName  :: Maybe Text -- ^ A unique name used to refer to this layer.
+  , _keycodes :: [Keycode]  -- ^ Layer settings containing also the buttons.
+  }
+  deriving (Show, Eq)
+makeClassy ''DefSrc
 
 -- | A mapping from names to button tokens
 type DefAlias = [(Text, DefButton)]
 
+data DefLayerSetting
+  = LSrcName Text
+  | LImplArnd ImplArnd
+  | LButton DefButton
+  deriving (Show, Eq)
+makeClassyPrisms ''DefLayerSetting
+
 -- | A layer of buttons
 data DefLayer = DefLayer
-  { _layerName :: Text        -- ^ A unique name used to refer to this layer
-  , _buttons   :: [DefButton] -- ^ A list of button tokens
+  { _layerName :: Text
+  , _layerSettings :: [DefLayerSetting]
   }
-  deriving Show
+  deriving (Show, Eq)
 
 
 --------------------------------------------------------------------------------
@@ -140,25 +171,26 @@ data IToken
   = KDeviceSource FilePath
   | KLowLevelHookSource
   | KIOKitSource (Maybe Text)
-  deriving Show
+  deriving (Show)
 
 -- | All different output-tokens KMonad can take
 data OToken
   = KUinputSink Text (Maybe Text)
   | KSendEventSink (Maybe (Int, Int))
   | KKextSink
-  deriving Show
+  deriving (Show)
 
 -- | All possible single settings
 data DefSetting
   = SIToken      IToken
   | SOToken      OToken
   | SCmpSeq      DefButton
-  | SInitStr     Text
   | SFallThrough Bool
   | SAllowCmd    Bool
   | SCmpSeqDelay Int
-  deriving Show
+  | SKeySeqDelay Int
+  | SImplArnd    ImplArnd
+  deriving (Show)
 makeClassyPrisms ''DefSetting
 
 -- | 'Eq' instance for a 'DefSetting'. Because every one of these options may be
@@ -168,9 +200,11 @@ instance Eq DefSetting where
   SIToken{}      == SIToken{}      = True
   SOToken{}      == SOToken{}      = True
   SCmpSeq{}      == SCmpSeq{}      = True
-  SInitStr{}     == SInitStr{}     = True
   SFallThrough{} == SFallThrough{} = True
   SAllowCmd{}    == SAllowCmd{}    = True
+  SImplArnd{}    == SImplArnd{}    = True
+  SCmpSeqDelay{} == SCmpSeqDelay{} = True
+  SKeySeqDelay{} == SKeySeqDelay{} = True
   _              == _              = False
 
 -- | A list of different 'DefSetting' values
@@ -185,7 +219,7 @@ data KExpr
   | KDefSrc   DefSrc
   | KDefLayer DefLayer
   | KDefAlias DefAlias
-  deriving Show
+  deriving (Show, Eq)
 makeClassyPrisms ''KExpr